% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyduck.R
\docType{package}
\name{tidyduck-package}
\alias{tidyduck-package}
\title{DuckDB high-level bindings for R}
\description{
R high-level bindings for DuckDB: an embeddable SQL OLAP Database Management System.
}
\seealso{
\code{\link[duckdb:duckdb]{duckdb::duckdb()}} for connection instructions.

\url{https://duckdb.org/} for the project website.
}
